chdir to the gem src when calling bundle open
This is relevant when using vim, vi, or MacVim with bundler.
You'll notice that if you do a bundle open and then try
to use PeepOpen, or even :edit, vim will think that you are still
in the directory that you called  from, rather than
the gem src, and it will think you want to open a file from your
project.

By doing a Dir.chdir() around the call to the editor this ensures
that we are in the right directory when vim is opened, so it's edit
commands work properly, then we will be returned to our project
directory by the time that bundle open is finished, so everything
works seamlessly.