change command impl doc to reflect dir structure of graph gem
add docs to recommend that gem command plugin implementors put their command
implementation in a file other than rubygems_plugin.rb that is loaded
by the command manager when it finds the constant missing.

Doing so makes it so that plugin writers do not have to ensure their plugin has not already
been loaded.