Pass basic auth details to the (possible) API endpoint. (Thanks dpiddy)
Given a source such as `http://foo:bar@gems.com`, the basic auth info
needs to be passed on so, if the source is not an API endpoint, it
will return 404 instead of 401. The fetcher will then fall back to
fetching the full specs as expected instead of bailing.

Conflicts:

	spec/install/gems/dependency_api_spec.rb