Unbreak gem specification loading if specification file encoding is not compatible with Encoding.default_internal
The :encoding option only specifies the external encoding.  If
Encoding.default_internal is set, it will automatically convert
it to the internal encoding.  If it cannot be converted (e.g.
internal encoding is US-ASCII and specification file contains
8-bit characters), an error is raised.

Instead, the :mode option should be given specifying that the
file should be left in its external encoding without converting
it to the default internal encoding.  This allows you to load
gem specification files with 8-bit characters and a default
internal encoding of US-ASCII.