Attempt to solve the DefaultKey problem once and for all
There are 4 yaml library scenarios that are now dealt with:
1) syck on 1.8
2) syck on early 1.9.2
3) syck on later 1.9.2
4) psych on >= 1.9.2

Cases 1 and 2 have syck loaded at YAML::Syck
Case 3 has syck loaded at ::Syck
Case 4 does not have a Syck constant at all

The code now detects and compensates for all 4 of these cases by making
sure that there is a YAML::Syck::DefaultKey and Syck::DefaultKey always.

Y::S::DK is needed to load yaml created in cases 1 and 2
S::DK is needed to load yaml created by case 3

In all cases, the code now prunes out the DefaultKey objects when the
yaml is loaded (#yaml_initialize time). This minimizes the exposure of
DefaultKey objects to the rest of the system.

If a DefaultKey object shows up misses the pruning code for some reason,
our DefaultKey#to_s is there to attempt to force ruby to ignore it and
output just a "=" like it should.