Make the Syck fixes resistent against YAML engine changes
If the YAML engine is changed after the syck_hack.rb file is loaded,
then things like the YAML toplevel constant are redefined to point to
either Syck or Psych. This therefore screws up being able to always
reference YAML::Syck::DefaultKey so on the load of syck_hack.rb, we get
everything the way we want it, then stash a reference to DefaultKey in
Gem so that we can always refer to it even if the YAML contstant is
redefined.