Extracted request into a method object.
This will open the door for much more change in the future.

More things could be pulled into this object, but I'm not
ready to fight with all the tests such things break yet.
Basically, #user_agent and the @proxy_uri variable are
only used within this method, but there are a bunch of
tests that hack at their internal state.