Gracefully handle remote source errors. Fixes #105
This is a little more aggressive than the pull request's diff because if
there is an error fetching the index for a source, it doesn't fall back
to the cache. This seems like the right behavior because if the index
can't be fetched, the likely good that the spec and .gem can be is very
slim.