Cleaned up class Date for modern ruby
Previously, require 'date' was expensive so we avoided it.  Now
Gem::Specification is loaded when needed so we don't have to worry about
this.

Removed FIX comment for TODAY constant as there is nothing to fix.  Time
is still preferred over Date as calculations are faster.  Two usages of
TODAY is enough to warrant a constant.  The local allows TODAY to be
correct for gems created at UTC midnight.

Fixed comment for Gem::Specification#date= that was attached to
DateTimeFormat.