Write gem entry checksums to a single entry
Previously checksums were written to an extra entry per entry in the gem.
(data.tar.gz -> data.tar.gz.sum which is a cost of 2N files per gem.)
Now checksums are written to a single checksums.yaml.gz as the last
entry in the gem.  (Cost of N + 1 files per gem.)

This reduces gem file size, especially if new files are added to the gem
format (a gem is a tar file, so the cost is at least 1024 bytes per file).

Additionally, the checksums are stored as a YAML document containing the
digest type and file name and digest contents which allows easier
parsing.  The previous format used a tab-separated list of digest
algorithms and digest values per line.