Support multiple digest algorithms for checksums
This allows for an upgrade process if further weaknesses in SHA1 are
found.  Currently SHA1 (used by signing) and SHA512 (strongest commonly
available) are used.  All checksums found in a file are verified rather
than giving a list of "best" checksums to check in order.

The digests extracted from a gem are now a Hash with an digest algorithm
containing a list of files and the hexdigest for that file and algorithm.

Gem::Package::TarReader#seek was added so checksums.yaml.gz can be
extracted before any other files are read.  This allows Package to more
easily verify the checksums embedded.

Gem::Package::DigestIO was extended to update multiple digests at once.