have a timeout around the gemcutter API endpoint
The API endpoint can sometimes hang and not return a response but open a
connection. Bundler before would just hang for a long time. This adds a
reasonable timeout, since the point of the API is to speed up the
install process. If the API takes too long, we should fall back on
cloudfront and the modern index, since that will likely be faster.