Fixed Policy#inspect
Previously inspect used a \ at the end of the string to continue to the
next line.  Due to a future change to Ruby this has been replaced with
a + to avoid warnings.

Replacing \ with + resulted in incorrect output as + has lower
precedence than % which would cause replacement to occur only for the
second part of the string.

Now the string is enclosed in parentheses to force proper order of
operations.