Always calculate package file digests
The updates for RubyGems 2.x introduced a checksums file for package
integrity.  Tar-format gems built with RubyGems 1.8.x and older do not
have these checksums.

While adding checksums the signing support was disabled for gems without
checksums.  Because the checksums did not exist the digest for signature
verification were not computed.

This commit fixes the package side of signature verification.  Now
digests are computed regardless of the presence of checksums.

A second bug in the verification code needs to check that the number of
signatures and digests match.