Refactor dummy signature generation
Now dummy signatures are all generated in dummy_signatures which can be
splatted onto a call to Policy#verify or #verify_signatures.  This makes
the code a little easier to scan and indicates the signatures should be
ignored as they aren't a particularly important part of the test.

This work is part of ensuring that, when verifying signatures,
signatures are present even if invalid.