Force signatures and digests when verifying
Previously if the user passed an empty digests array verification would
succeed.

This could lead to lack of verification when Package has a bug (such as
54dc22f).

Now digests must be filled and the number of signatures must match the
number of digests (only the reverse was tested previously) to help catch
such bugs.

This check is performed even when data integrity is disabled to warn the
user of tampering.