Properly test --install-dir behavior
The --install-dir option for gem install is supposed to create a
stand-alone installation of the listed gems.  It is supposed to ignore
gems already installed in the gem home or gem path.

The test for this feature did not test installing dependencies into the
chosen install directory.

These tests have been added.

Closes #442