Fixed placement of executables with --user-install
Previously executables for --user-install were placed in ~/.gem.  It
seems this broke at some time during the 1.8.x series due to a missing
test.

This caused an attempt to install executables into the default bin
directory which is usually only writable by a super-user.

This was fixed by setting the bin directory to the appropriate path if
the :user_install option was enabled.