Create extension destination directory
After an extconf.rb extension is built it is installed into the "lib"
directory.  If the "lib" directory was not present in the package this
would cause the shared object to be named "lib" instead of its proper
name.

This patch creates the destination directory before building the
extension to ensure the shared object will be installed with the correct
name.

Ruby Bug #7897, patch by Kenta Murata