Lazily load RubyGems.arg API credentials
Eagerly loading the credentials causes the .gem/credentials security
check to occur even when the API credentials will not be used.  This can
cause a failure to install when running setup.rb.

Now the credentials file is lazily loaded which will delay the check
until the credentials are used (such as pushing a gem).  This allows
users to install rubygems through automated mechanisms (such as RVM)
with bad permissions on their credentials file.

Fixes #465