Begin move to DependencyResolver for installation
This is a work-in-progress that, at least, doesn't give errors for most
tests.  Due to incomplete integration many DependencyInstaller tests
still fail due to not installing enough gems.

A new resolver set was added, Gem::DependencyResolver::InstallerSet,
which handles both local and remote gems.  This should either be merged
with IndexSet or contain an IndexSet for remote operation.

Also, a typo in UnsatisfiableDependencyError was corrected.