Fix --ignore-dependencies
Previously the current set of installed gems was used.  This created
incorrect behavior if the gem was already installed.

Now we rely on the always_install list and return no dependencies unless
the gem already exists in the always_install list.  This means the gem
will always be installed (as it was listed).