Match --minimal-deps test to its description
`gem help install` says "Don't upgrade any dependencies that already
meet version requirements" for --minimal-deps.  However, the test does
not provide an option to upgrade and stops once a direct dependency has
been met, even if the set is incomplete.

Now we traverse all the dependencies and have a newer version of a
dependency available to install.  This ensures a complete set is
installed.