Convert version to requirement when updating gems
Gem::DependencyInstaller expects a Gem::Requirement but the updated
sends a Gem::Version instead, so convert it.  While #570 describes the
issue and provides a backtrace it does not provide enough information to
reproduce the issue locally.  Perhaps a local version of the gem is
required to trigger the bug?

Fixes #570