Reset RUBYOPT when unlinking siteconf
The rice gem uses an unconventional extconf.rb during its build.
Instead of using mkmf it gathers configuration information from RbConfig
and runs a configure script which creates a Makefile.  The Makefile
invokes ruby during the build process.

When running the Makefile the RUBYOPT environment variable required the
siteconf file which had been removed.  This would cause ruby to raise an
exception.

Now RUBYOPT is reset when the siteconf file is removed which fixes this
bug.

Fixes #582