Fix data.tar.gz checksum in test_add_checksums
The expectation rebuilt data.tar.gz to determine the checksum.  If the
original and the expected tar files were built in different seconds the
test would fail as a timestamp is embedded in the gzip footer.

Now the tar.gz is extracted from the gem and the checksum is computed
from a binary-identical file instead of a content-identical file.

Fixes bug found by Koichi Sasada.