Replace RAKE_SUCKS workaround for hoe's docs task
The RAKE_SUCKS workaround was a fragile hack dealing with the
inflexibility of Rake::TaskLib combined with the inflexibility of hoe.
The workaround eventually broke and would always fail to generate
correct documentation.

Now hoe's rake tasks are removed and replaced with an RDoc::Task created
using information from hoe.