Place the remote source first when sorting
The dependency resolver prefers the last source to use for installing a
gem as it will be the source containing the highest version for remote
gems.  Local gems (installed or as .gem files) need to be higher because
the user is requesting those specifically.

This prefers a named local gem over a remote gem with the same version
which saves a fetch and allows use of a modified and repackaged local
gem.

Fixes #616