Fixing Gem.clear_paths to work in an environment with a Security Constant
Sometimes rubygems will execute in an environment where there is a top level Security constant.
Checking for const_defined? :Security would return true when this Security constant exists, even
if it's not the expected Gem::Security.

defined? does not have this problem, and allows us to check for Gem::Security directly.