Include platform when sorting gems for resolution
When there are multiple possible gems to use for resolution RubyGems
would sort by source (name) and version and choose the latest spec.
This did not include the platform.

Omitting the platform could result in a gem with a C extension being
chosen over the platform-specific version.

Now the various specification classes (in the resolver) include the
platform and the resolver includes platform when sorting.  This causes
the resolver to try using a platform gem before using a gem with a C
extension.

Fixes #635