Use platform to load specs in DependencyResolver
Previously the platform was ignored when fetching specs inside the
dependency resolver.  This meant if gems with different platforms had
the same name and version, a spec with the incorrect platform could be
selected and installed.

Now the platform is included when fetching the spec which forces loading
of the correct specification.

Fixes #635 (again)