Fix installation of prerelease gems
When available_specs was introduced it didn't sort the :complete type
(which contains sources from two datasets).

Binary search was added to speed up extracting only the needed tuples
from the result of available_specs (see #595).  Binary search depends on
a sorted list.

The assumption that available_specs returns a sorted list was masked by
the tests which sorted the expected value.  The other available_specs
results are all sorted by default, so this was an easy one to miss on
review.

Fixes #636