Fix installing extensions with --install-dir
When the extension building code was moved from Gem::Installer to
Gem::Ext::Builder the dependency on gem_dir was not correctly migrated.

This meant that the gem_dir computed in Gem::Ext::Builder was incorrect
so RubyGems would attempt to install the extension in GEM_HOME not the
user's chosen directory.

Issue #642