Fix installation of gems with "./" files
In the files list of some gems (contained in the specification and in
the data.tar.gz) the entries start with "./".  In RubyGems 2.1.0 through
2.1.2 such gems install, but none of their files appear in the gem dir.

Commit @5be6784 introduced support for installing gems as default gems
directly from RubyGems.  This added a File.fnmatch filter to extracting
the files in a specification.

File.fnmatch '*' (the default) does not match file entries starting with
"./", resulting in the bug described above.

I'm unsure of the source of the "./" in the file lists or tar.gz.

Fixes #644

Fixes #645

Fixes #646