Allow `gem uninstall foo --all`
`gem uninstall --all` was not intended to work with gem names, but this
may introduce a regression (I did not check, but #650 says this is true)
when force-uninstalling all versions of a gem.

Now `--all` can be used with a gem name to force-uninstall all versions
of that gems without questions.

Fixes #650