Only display relevant release notes upon update
Commit @968626d introduced a --previous-version option for updating
RubyGems to control display of history information, but took no
advantage of it.

Now, if the user is updating to a version that supports the option,
RubyGems will send it to the new version so it displays only the
relevant lines from the history file.

Fixes #647