Don't set --install-dir in the uninstall command
Way back in @e4bb356 I listed --install-dir as a default option to the
uninstall command even though it was not a default option.

In @1b6f55b I added --install-dir as a default option for the uninstall
command because it was listed in the default arguments, thinking this
was simply an oversight.

This broke the `--user-install` by default behavior of `gem uninstall`.

This commit partially reverts @1b6f55b

Fixes #653