Reduce specificity of install command tests
The install command tests would test exact output, but due to
interactions between the install command and various plugins it is
difficult to ensure the output is exact if the test order changes.

Now, instead of assert_equal on a particular line, assert_match is used
to ensure that certain important information is displayed.