Allow Gem::Ext::Builder to accept only the spec
At install time the build_info file is not yet present so
Gem::Specification#build_args will be empty (this was an oversight that
should be corrected separately, but backwards compatibility must be
maintained).  This meant that the build_args needed to be passed
separately.

After install time the build_args will be present so we can let the
builder fetch them from the specification directly.