Rename extension build error constant
The constant Gem::Installer::ExtensionBuildError required loading the
installer at extension build time which could be a large amount of
files.  This constant is now an alias of Gem::Ext::BuildError.

The new constant is needed to minimize loading when extensions are built
during activation.

The old constant is deprecated and nodoc'd.