Include extensions in stub specification data
This will allow require_paths and full_require_paths to be pushed up to
BasicSpecification where they can be shared between the two
specification types.

This requires a second stub line as the final value in the stub
specification is the require paths which may include spaces.  This means
the extensions must go on another line as the spaces are not escaped nor
can we change the format of the stub line.