Move require_paths up to BasicSpecification
In order to determine if a gem contains a particular requirable file
the BasicSpecification also needs to look in the extension install
directory.

This means the two methods (require_paths and extension_install_dir) to
support this need to move up to Gem::BasicSpecification from
Gem::Specification.

This commit brings up a question:  Does building the extensions need to
happen sooner than activate?  In contains_requirable_file? instead?