Move extension building to file existence check
Previously extension building occurred on activate, but this was wrong.
If a gem like nokogiri was installed for a different platform it could
never activate because the extension would always be missing during the
file existence check (contains_requirable_file?).

Now, extension building happens during the existence check.

This moved gem_dir up to BasicSpecification and causes the stub
specification to load the real specification to build the extension.