Always write gem_make.out
If there is an extension build that reports success but was not
successful it would be nice to see what occurred during the build logs.

Previously this was not possible as the gem_make.out file was not
written when extension building reported success.

Now it is written every time an extension is built.  The existence of a
.gem.build_complete file can be used to check for success instead, this
file is deleted before retrying the build.