Use string platform for resolver specifications
When looking up specifications the resolver sets expect string
platforms.

For a local gem the parsed platform was available and used for looking
up the specification.  The specification was stored with a string
platform which meant the gem was not available (causing an exception).

Now the IndexSpecification always uses the string version of the
platform to prevent this bug.

Fixes #664