Enhance extension failure reporting
Previously if an exception was raised the exception would not be
reported.  Additionally, if a command failed a large message would be
built but thrown away as soon as build_error was reached.

Now a small exception message is raised for the original error and this
message included in the build report.

This will make determining build failures easier.