Re-raise exceptions when looking for latest gems
In the update command RubyGems was ignoring exceptions raised when
determining the latest version.

This meant that `gem update --system` from a rubygems with invalid
certificates would incorrectly report "nothing to update".

Now, if an error is raised and captured, the update command will
re-raise the exception which gives useful feedback to the user.

Fixes #688