Disallow gem activation while building extensions
RubyGems needs to require files in $LOAD_PATH to build an extension
while it is determining which gems to load.  This is not possible if
there are unresolved dependencies as rubygems will attempt to load those
dependencies which need to build extensions, which creates an infinite
loop (that is fortunately stopped by the CHDIR_MUTEX while building
extensions.

Now RubyGems temporarily clears Gem::Specification.unresolved_deps while
it requires the necessary files, then restores it after the extension is
built.  This is OK to do as the RUBYGEMS_ACTIVATION_MONITOR is held in
the overridden Kernel#require.