Fix updating gems which have multiple platforms
After @0000d57 for #688 was committed, when updating a gem with multiple
platforms a TypeError would be raised.  This occurred because a
PlatformMismatch error was returned which we don't care about in the
update command.

The PlatformMismatch is not an Exception so it could not be raised.  Now
only errors that respond to #exception are raised, which is a poor way
of ignoring the PlatformMismatch errors inside fetch_remote_gems.

Fixes #693