Try all states when resolving dependencies
DependencyResolver#find_conflict_state chose only the state with a
matching dependency and ignored all other states.

This meant that if there was another state that would result in
successful resolution of the states the resolver would never try it.

Now find_conflict_state prefers selecting a state with a matching
dependency, but returns the first state if that did not work out.  This
allows the resolver to backtrack successfully.

@evanphx should review this commit.