Add stubbed? method to Gem::BasicSpecification and Gem::Specification
I originally thought this did not make sense to do, since a question
about whether the spec is stubbed or not only exists if the stub is a
Gem::StubSpecification.

However, I have discovered that Bundler overwrite
Gem::Specification.stubs, so that it returns Gem::Specification
instances. Therefore I think it is useful to define this method on
Gem::Specification for the purpose of duck typing.