Set Gem.sources directly for source in gem.deps.rb
There isn't a provision for setting sources for a particular resolve set
at this time.  This means we can't override the sources for just this
one resolution.

Instead, we set the sources directly on Gem.sources and use them.  Since
multiple gem dependencies files are typically not used in the same
process this should not be an issue for time being.