Allow correct activation of path: gems
Previously RubyGems assumed that a path: gem lived in the same structure
as a regular gem.  This would mean use of a gem.deps.rb would not work
correctly as the files in the gem would not be accessible through
$LOAD_PATH.

Now the root of the gem's files can be set on a specification to allow
correct operation.