Use the .lock file to install gem dependencies
Now a .lock file will be used if present when installing gem
dependencies.  This is designed to ensure the same gem versions are used
across multiple installs.

I'm not sure if RubyGems' resolver architecture behaves identically to
bundler.  RubyGems doesn't presently run bundler's integration tests, so
compatibility is not known.