Restore installing remote gem without dependencies
Commit @d97fba1 was intended to speed up the bundler tests but
inadvertently removed the ability to install remote gems.

The DependencyInstaller normally figures out where gems live if only a
name is given, but @d97fba1 bypasses that code, so only local gems with
a full path are recognized.

The lookup code from DependencyInstaller was reimplemented in
InstallCommand to restore the feature.