Reduce HTTP requests by using BestSet
RubyGems 2.2.0 introduced use of the bundler API, but accidentally this
was not complete.  `gem install` of a gem was left using the marshal API
like 2.1.x and earlier which makes lots of requests.

Now RubyGems uses BestSet inside InstallerSet for remote requests.  This
should speed up install during the dependency fetching phase.

This should also address #762