Adding CONTRIBUTING file.
There is some info on the [RubyGems Site][1] that wasn't easy to find in
this repo. In an effort to make it easier for others to start contributing, as
well as taking advantage of a nice [CONTRIBUTING FILE SUPPORT][2] on Github I'm
copying that information here for people thinking of contributing to RubyGems.

[1]: http://guides.rubygems.org/contributing/
[2]: https://github.com/blog/1184-contributing-guidelines