Support multiple sources in gem.deps.rb.lock
The GEMS section lists all sources for all gems and does not pin gems to
a particular source.  This was missing, so now we add all gems using
each source for lookup.

This commit is suspect.  If the dependency API is not available for a
source what happens?  See future commits.

See #851