Disable dev dependency checking for install_into
Installer does not know the difference between shallow and deep
development dependencies and raises an exception when set in development
mode for deep dependencies.

RequestSet has already done the work of checking the dependencies so the
extra check in Installer is redundant and wrong.

Instead of fixing Installer instantiation to understand the difference
for this installation (as the information has been lost) we just disable
development mode as we can trust RequestSet.

Part of #893