Update Resolver::Specification#spec after install
This allows documentation generation and other hooks to work properly
after install.  Before the path was incorrect so documentation would try
to generate into the default installation directory which is sometimes
wrong (like with -i).

Following this I need to provide a @spec accessor at top-level because
everybody seems to use one, along with an audit of caching in
subclasses.