Raise ArgumentError when gem deps file is missing
Now that Gem.use_gemdeps prefers an argument it should raise an
exception if the file given should not be found.

Since the method falls back to the RUBYGEMS_GEMDEPS environment variable
(which is used via gem executable stubs) the method does not raise an
exception when the argument matches the RUBYGEMS_GEMDEPS environment
variable.

This is not great because sometimes an exception will not be raised
which will be confusing, but it does maintain backwards compatibility.