Ignore ruby in gem.deps.rb when installing
Now RubyGems allows a gem dependencies file to be used for installing
(but not for activation) when the ruby version does not match the one
specified in the gem.deps.rb.

Previously RubyGems would raise a VersionMismatch exception even when
installing.  This seems less-useful than allowing installation with the
version mismatch.

RubyGems may install incorrect gems in this case if there are
ruby-version-specific gems that are only installable on the ruby
version listed in the gem.deps.rb.

Fixes #941

Pair: @Elffers