Add a "whereis" command.
"gem whereis GEMNAME [REQUIREMENT ...]" shows the base directory where a
specified gem is installed.  This is considered useful for use in
scripts and editor plugins, or when one simply wants to browse a gem's
installed directory.

"gem which" does a similar job, but it is focused on locating a library
file and does not work for gems that do not provide a user loadable
library.